// src/types/menu.ts
export enum MenuItemCategory {
    BREAKFAST = 'Breakfast',
    LUNCH = 'Lunch',
    DESSERT = 'Dessert',
    HOT_BEVERAGE = 'Hot Beverages',
    COLD_BEVERAGE = 'Cold Beverages',
    DRINK = 'Drinks',
    EXTRA = 'Extras',
    LUMERE_FLORAL_SPECIALS = 'Lumere Floral Specials',
    LUMERE_SPECIAL = 'Lumere Special'
}